third_party/blink/web_tests/platform/generic/fast/forms/number/number-spinbutton-changeevent-trigger-expected.txt

Should not trigger any event
PASS inputEventCounter is 0
PASS changeEventCounter is 0
Input event is triggered but no change event is triggered on mouse down
PASS inputEventCounter is 1
PASS changeEventCounter is 0
PASS numberInput1.value is "1"
Change events occurs on mouse up
PASS inputEventCounter is 1
PASS changeEventCounter is 1
PASS numberInput1.value is "1"
Pressed more than once, still events are triggered
PASS inputEventCounter is 2
PASS changeEventCounter is 2
PASS numberInput1.value is "2"
When moving away from field, it should not trigger event, only on value change
PASS inputEventCounter is 2
PASS changeEventCounter is 2
PASS numberInput1.value is "2"
PASS successfullyParsed is true

TEST COMPLETE
